python/requirements.txt
absl-py==2.3.1
aistudio_sdk==0.3.5
annotated-types==0.7.0
anyio==4.10.0
arabic-reshaper==3.0.0
astunparse==1.6.3
bce-python-sdk==0.9.42
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
easyocr==1.7.2
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.7.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
Jinja2==3.1.6
keras==3.11.3
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
modelscope==1.29.1
mpmath==1.3.0
namex==0.1.0
nepali-datetime==1.0.8.4
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
optree==0.17.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.0
pandas==2.3.2
pillow==11.3.0
prettytable==3.16.0
protobuf==6.32.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.2.3
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.0
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tifffile==2025.8.28
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
PyPDF4==1.27.0
cryptography==43.0.1
reportlab==4.2.5
pdf2image==1.17.0
